Core functionalities implemented.
Now the library can do the job. It can fetch the migration files and execute the queries on the database and record the applied migrations into specific table called migration_history. it checks the checksum of the files to prevent modification of the migration after it's applied. It is idempotent. And much more.